+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.0.1.dev
 - Added `ngram` function
+- Added `is_empty` and `is_nonempty` functions
 
 ## 2.0.0 (March 19, 2018)